add comments

+++ hashTable/fileProcessing.h
@@ -2,4 +2,4 @@
 #include "hashTable.h"
 
 // creates hash table and fills it with frequencies from file
-HashTable *getFrequenciesFromFile(int hashTableSize, const char *filePath, int *errorCode);
+HashTable *getFrequenciesFromFile(size_t hashTableSize, const char *filePath, int *errorCode);
